modify code_file_name
