--- Week 1/id_155/66.加一.java
@@ -1,26 +1,0 @@
-import java.util.LinkedList;
-
-/*
- * @lc app=leetcode.cn id=66 lang=java
- *
- * [66] 加一
- */
-
-// @lc code=start
-class Solution {
-    public int[] plusOne(int[] digits) {
-        for(int i = digits.length - 1; i >= 0; i--){
-            digits[i]++;
-            digits[i] = digits[i] % 10;
-            if(digits[i] != 0){ // 没有进位 直接返回，有进位也只可能进 1
-                return digits;
-            }
-        }
-        // 数组一直在进位 eg：999 -> 1000
-        digits = new int[digits.length + 1];
-        digits[0] = 1;
-        return digits;
-    }
-}
-// @lc code=end
-
